Changing ctest list designation for some cases.  There are currently
57 Sanity tests
32 Checkin tests
33 Continuous tests
11 Nightly tests
and
11 Occasional tests

The timeout criteria for the different cases is set to
30 sec. for Sanity tests
100 sec. for Checkin tests
600 sec. for Continuous tests
2000 sec. for Nightly tests
7200 sec. for Occasional tests

The ctest output from an 8 processor run of all tests except those in the
Occasional category completes in less than an hour on my mac with the following
output.

100% tests passed, 0 tests failed out of 133

Label Time Summary:
Checkin       = 254.00 sec
Continuous    = 1238.00 sec
Nightly       = 2855.00 sec
Sanity        =  77.00 sec


This should speed up the test harness and also eliminate most or all of the
timeouts we have been observing on various platforms.

LF