Updated the code to intialize the data in the solution vector to 0 before passing it to MueLu.
It looks like MueLu is having the same issue as the Tramonto code with
propagating NaNs. Belos is calling the preconditioner on uninitialized
"Cloned" vectors, expecting the underlying operators to have overwrite
semantics. We should ask Jonathan what the expected behavior for MueLu
should be. Right now when used with Belos, you have to initialize the
solution vector to 0 to avoid NaNs before passing it to the MueLu
Hierarchy and calling Iterate().

This fixes the POLY26, POLY27, POLY28, and DIFFUSION2 examples for me,
including the NaN problem when running on 8 processors. The other
examples here were simply failing and not reporting NaNs.