I updated several important settings to make the Tpetra code more closely emulate the Epetra code, which speeds up the test harness for Tpetra:
1. Changed the PolyA22_Coulomb_Tpetra_Operator class to simply apply the
MueLu hierarchy as a preconditioner, NOT use it in any iterations (by
calling Iterate()). This speeds things up quite a bit.

2. Manually set the fill-in and drop parameters for the Ifpack2 preconditioner
for the A22 block to the same settings that Epetra is using. I was using more
aggressive settins before, but this slows things down. I haven't tested this
yet on all the harder examples.

3. Setup a separate Belos list to use in the TpetraPolyLinProbMgr and the
TpetraHardSphereProbMgr in the Tpetra_ParameterListConverter class. The Epetra
code currently manually sets the number of iterations and kspace size to 500
in these classes, instead of using inputs from the physics code, which may make
sense. However, this slowed down the examples for me so I haven't enabled these
settings yet.