Removed an unused type, and generalized the Tpetra_MultiVector class and Tpetra_MixedOperator classes to work with any two precisions, not just Scalar and halfScalar.
The Tpetra_MultiVector class is templated on the
DomainScalar and RangeScalar now, and converts between the two types. The
Tpetra_MixedOperator class is templated on the Scalar type of the overall
operator, and on the OpScalar type of the underlying operator. It uses
MultiVector converter to convert the input/output multivectors.