1. Added back in the BelosTypes.hpp header beacuse it's needed for optionally printing Belos details.
2. Now using more aggressive level-of-fill (7 instead of 2) for the A22 block
when A22 is a matrix for HardSphere problems. This could cause a slow down for
other examples, although the test harness is running at roughly the same speed
for me now.

One performance problem is that Ifpack2::AdditiveSchwarz really needs
to split off more of the work that pertains only to the graph structure from
the work that relies on the matrix values. Currently, to function correctly we
need to call initialize() and compute() every time (even though only the matrix
values have changed), thus redoing a bunch of the work on the matrix structure.
The Ifpack2::Diagonal preconditioner also could benefit from similar optimizations like
caching the locations of the matrix diagonals instead of doing expensive searches.
This should be easier to address than the problem with Additive Schwarz so I'll
likely get to it sooner.