Applying patch from Chetan Jhurani to correct a file path length issue
0002-change-FILENAME_LENGTH-to-4096.patch

We discovered that when using the GUI, the file path length was limited
and that causes problems in some very realistic setups.

I ran the checkin tests with the CLANG compiler, and the Sanity tests
with the GCC 4.8 compiler prior to committing (a few checkin tests fail
using the GCC 4.8 compiler on my machine prior to applying the patch).